pyproject.toml
[project]
name = "livekit-twilio-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "livekit-agents[aws,deepgram,google,silero,turn-detector]~=1.2",
    "pydantic-settings>=2.12.0",
]
